cli/requirements.txt
sentence-transformers>=3.0.0
umap-learn>=0.5.6
numpy>=1.24.0
